Delete order feature
A feature to delete an order was implemented, it's only available to admins and only deletes the order by the id provided if the order has a Delivered or Canceled status (meaning it's completed). The on delete cascade option was added to the order reference in the order products schema, since if an order is deleted there is no longer any need for the order products.